requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.11
matplotlib>=3.7
seaborn>=0.13
torch>=2.0
torchvision>=0.15
transformers>=4.33
datasets>=2.14
pyyaml>=6.0
pytest>=7.0
